+++ cmake_format/tests.py
@@ -3,10 +3,11 @@ import unittest
 # pylint: disable=wildcard-import
 # pylint: disable=unused-wildcard-import
 
+from cmake_format.format_tests import *
+from cmake_format.invocation_tests import *
 from cmake_format.lexer_tests import *
 from cmake_format.markup_tests import *
 from cmake_format.parser_tests import *
-from cmake_format.format_tests import *
 
 if __name__ == '__main__':
   unittest.main()